feat: delete account

+++ app/Http/Controllers/Users/DeleteUserController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Users;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\CreateDraftRequest;
 use App\Http\Requests\CreateUserRequest;
+use App\Http\Requests\DeleteUserRequest;
 use App\Models\Draft;
 use App\Models\DraftPick;
 use App\Models\User;
@@ -14,13 +15,15 @@ use Illuminate\Http\Request;
 
 class DeleteUserController extends Controller
 {
-    public function __invoke(Request $request)
+    public function __invoke(DeleteUserRequest $request)
     {
         $data = $request->validated();
 
         $user = User::where('external_id', $data['external_id'])->first();
 
-        $user->fantasyTeams()->detach();
+        if (!$user) {
+            return response()->json('User not found', 404);
+        }
 
         $user->delete();
